requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
gunicorn==21.2.0